learner/requirements.txt
allure-python-commons==2.2.2b2
asn1crypto==0.24.0
attrdict==2.0.0
attrs==17.2.0
Automat==0.6.0
backports.shutil-get-terminal-size==1.0.0
basicstruct==1.0.3
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.7
coloredlogs==10.0
constantly==15.1.0
cryptography==2.3.1
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
Distance==0.1.3
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.11
django-appconf==0.6
django-compressor==1.4
django-filter==0.15.2
django-filters==0.2.1
djangorestframework==3.4.7
docopt==0.6.2
drf-nested-routers==0.9.0
-e git+https://github.com/amir9979/Debugger.git@e93cc532e1a7277bae584ae13a7d1e961bb8663b#egg=eclipse&subdirectory=learner
enum34==1.1.6
Flask==0.12.2
functools32==3.2.3.post2
future==0.17.1
gitdb==0.6.4
gitdb2==2.0.5
github3.py==1.2.0
GitPython==2.1.11
gunicorn==19.1.1
humanfriendly==4.17
hyperlink==17.3.1
idna==2.7
ijson==2.3
incremental==17.5.0
ipaddress==1.0.22
ipdb==0.10.1
ipdbugger==1.1.2
ipython==5.0.0
ipython-genutils==0.1.0
itsdangerous==0.24
javalang==0.12.0
Jinja2==2.9.6
jira==2.0.0
jsonschema==2.6.0
junitparser==1.2.2
jwcrypto==0.6.0
liac-arff==2.3.1
lxml==4.0.0
Markdown==2.6.7
MarkupSafe==1.0
matplotlib==2.0.0
monotonic==1.5
namedlist==1.7
networkx==2.2
nose==1.3.7
npm==0.1.1
numpy==1.12.0
oauthlib==2.1.0
optional-django==0.1.0
pandas==0.19.2
pathlib2==2.1.0
pbr==5.1.1
pickleshare==0.7.3
Pillow==3.3.0
pluggy==0.5.2
prompt-toolkit==1.0.3
psutil==5.3.1
psycopg2==2.6.2
py==1.4.34
pycparser==2.19
Pygments==2.1.3
pyjwt==1.6.4
pymongo==3.4.0
pyparsing==2.2.0
pyreadline==2.1
pyswarm==0.6
pytest==3.2.3
pytest-allure-adaptor==1.7.8
python-bugzilla==2.2.0
python-dateutil==2.6.0
pytz==2016.10
requests==2.20.1
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
rotest==2.4.0
scipy==1.1.0
selenium==2.53.6
-e git+https://github.com/amir9979/SFL-diagnoser.git@51d5d7788030a00635b24b139d97f2d1bbddc060#egg=sfl_diagnoser
simplegeneric==0.8.1
six==1.10.0
smmap==0.9.0
smmap2==2.0.5
static3==0.5.1
termcolor==1.1.0
traitlets==4.2.2
Twisted==17.9.0
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==15.0.2
wcwidth==0.1.7
Werkzeug==0.12.2
win-unicode-console==0.5
WTForms==2.1
xlwt==1.3.0
xmltodict==0.11.0
zope.interface==4.4.3
